make bar and histogram have distinct defaults step:
- split bar style defaults logic into own file to be reused
  by histogram defaults
- split histogram bin defaults logic into own file to be reused
  by histogram2d and histogram2dcontour defaults
